SCRUM-91 feat: implement terms of service page
